sửa lỗi double

+++ BlindBox Backend/src/main/java/com/blindbox/controller/UserAccountController.java
@@ -4,8 +4,6 @@ import com.blindbox.service.UserAccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
-
 @RestController
 @RequestMapping("/user-account")
 public class UserAccountController {
@@ -13,9 +11,9 @@ public class UserAccountController {
     @Autowired
     private UserAccountService userAccountService;
 
-
+    // API để tạo tài khoản người dùng với initialBalance kiểu double
     @PostMapping("/create")
-    public String createUserAccount(@RequestParam Integer userId, @RequestParam BigDecimal initialBalance) {
+    public String createUserAccount(@RequestParam Integer userId, @RequestParam double initialBalance) {
         userAccountService.createUserAccount(userId, initialBalance);
         return "User account created successfully!";
     }
